Migration to golang.org/x/sync/errgroup.
