+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/mabrarov/mp
 
-go 1.23
+go 1.23.0
+
+require golang.org/x/sync v0.16.0
