--- pkg/supervisor/process.go
@@ -1,5 +1,0 @@
-package supervisor
-
-import "context"
-
-type Process func(ctx context.Context) error
