--- pkg/supervisor/supervisor.go
@@ -1,75 +1,0 @@
-package supervisor
-
-import (
-	"context"
-	"errors"
-	"sync"
-
-	"github.com/mabrarov/mp/pkg/panicerr"
-)
-
-type Supervisor struct {
-	ctx    context.Context
-	cancel context.CancelFunc
-	group  sync.WaitGroup
-	mutex  sync.Mutex
-	err    error
-}
-
-func New(ctx context.Context) *Supervisor {
-	ctx, cancel := context.WithCancel(ctx)
-	return &Supervisor{
-		ctx:    ctx,
-		cancel: cancel,
-		group:  sync.WaitGroup{},
-		mutex:  sync.Mutex{},
-		err:    nil,
-	}
-}
-
-func (s *Supervisor) Go(p Process) {
-	s.group.Add(1)
-	go func() {
-		defer s.group.Done()
-		s.setError(s.run(p))
-	}()
-}
-
-func (s *Supervisor) Wait() error {
-	s.group.Wait()
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-	return s.err
-}
-
-func (s *Supervisor) Stop() {
-	s.cancel()
-}
-
-func (s *Supervisor) run(p Process) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			pe := panicerr.New(r)
-			if err == nil {
-				err = pe
-			} else {
-				err = errors.Join(err, pe)
-			}
-			s.Stop()
-		}
-	}()
-	return p(s.ctx)
-}
-
-func (s *Supervisor) setError(err error) {
-	if err == nil {
-		return
-	}
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-	if s.err == nil {
-		s.err = err
-	} else {
-		s.err = errors.Join(s.err, err)
-	}
-}
